Temporarily disable the multi-query test for ulib-postgres (#4058)
In the past couple of continuous benchmarking runs, ulib-postgres moved
into first place in the multi-query test by a wide margin while being
much further down the list in the single query test.  This is symptomatic
of batch queries, which are disallowed.

As far as I can tell by looking at the implementation and framework, the
same batching logic that I rejected in earlier PRs is still there, except
the batching logic was moved into the framework itself.

We can re-enable these tests for ulib later if someone convinces us that
it's not doing batch queries.