Add Json benchmark support to aspnetcore-corert (#4212)
* Add Json benchmark support to aspnetcore-corert

This is pretty much just a copypaste of the Json parts of the aspnetcore benchmark.

The notable difference is the Utf8Json.Generated.cs file - this file was generated using the Utf8Json.UniversalCodeGenerator tool provided by the Utf8Json author. Utf8Json relies on Reflection.Emit by default (and the presence of a JIT). The Reflection.Emit part wouldn't work on CoreRT.

* Update README.md

* Add header to Utf8Json.Generated.cs