H2O: Prevent usage of discarded HTTP request structures
The client may close the connection or discard the HTTP/2 stream
while the application is waiting for a result from the database,
in which case the h2o_req_t structure will be deallocated. As a
result, the database callbacks will subsequently try to access
memory they are not supposed to. The h2o_mem_alloc_shared()
function provides a mechanism to receive a notification when such
a situation occurs.