Update roda-sequel to use correct block arity (#4615)
The static_get blocks are supposed to accept a request
argument.  They have always been yielded a request argument,
it has just been ignored by this code (Ruby blocks are not
strict in regards to arity).

Starting in the most recent release of Roda, there is a
significant performance improvement by have the block
accept an argument.