Upload results to tfb-status after each failure (#5289)
This should fix an issue we're seeing with false positive 'Benchmarking
environment "Citrine" crashed?' emails internally.  When one framework
has many variants that are failing in a row, and each one of them is
taking an hour to fail, there may be several hours between successful
tests.  Currently this means several hours between uploads, and this
triggers tfb-status's "crashed?" email.  With this change, there should
be not much more than an hour between uploads at worst.

Also added what looked like a missing call to write_intermediate, which
uploads the "completed" key in the results.json, in the case of a database
failure.