Update Python - Quart framework (#5527)
* Upgrade Quart requirements

This upgrades to the latest versions. For benchmarking this should be
useful as Hypercorn has improved since 0.5.3.

* Minor updates to Quart app file

The DB should be initialised in the before_serving i.e. before any
request is received, rather than the before_first_request (on first
request). This is now possible with Hypercorn and Uvicorn (likely
wasn't when code was first written).

Also routes should be async (or they are wrapped in threads with a
performance hit).

* Upgrade Quart-Uvicorn requirements

This upgrades to the latest versions. For benchmarking this should be
useful as Uvicorn has improved.

* Disconnect from DB in Quart app

May not affect benchmarks, unless there are cleanup issues.