Yii2: enable the ORM cache on the DB schema (#5541)
* yii2: replace PHP 7.3 references by 7.4

* yii2 fortunes: sorting with anonymous function was slower

My micro-bencharks showed that using an anonymous function
as a callable parameter was about twice slower than using
a named function or a static method.

The gain is minor but visible, about 0.5 µs/call,
so theorically about 300 µs/request.

* yii2 updates: skip model validation

No visible gain, but it makes sense to skip the call to
Fortune::validate() when the data is not an user input.

* yii2: cache the DB schema

Without this, Yii2 analyzes the DB schema at the beginning of
each request. This caused warnings in `tfb` about the surpringly
large number of SQL queries.

The gain is huge. For instance, in yii2-fortunes it was more than ×7,
from 387 req/s to 2824 req/s on my system.

* yii2: do not install Bootstrap CSS components

Minimal performance gain (less classes for the autoloader),
so this change is mostly a sensible cleanup.

Co-authored-by: François Gannaz <francois.gannaz@silecs.info>