Temporarily disable ffead (#5815)
The last two runs in the Citrine environment failed because of
problems induced by ffead.  We don't know that ffead is doing
anything wrong -- even if it is, the TFB toolset is to blame for
letting that ruin the run -- but we would like to continue
benchmarking other frameworks while we figure this out.