Python: Fix asyncpg PreparedStatement usage
asyncpg's PreparedStatement
(https://magicstack.github.io/asyncpg/current/api/index.html#prepared-statements)
will cache results in a local LRU cache which is not desired as per
https://github.com/TechEmpower/FrameworkBenchmarks/wiki/Project-Information-Framework-Tests-Overview#general-test-requirements
this results in some test failures whereby not enough queries are
completed.

The fix is to sample the range and hence get the `queries` number of
unique IDs rather than potentially get overlapping IDs.

I've done this for the frameworks I am familiar enough with to change
with confidence. It may apply to the other frameworks using asyncpg.