Updates & fixes to Python frameworks (#7311)
* Python-Quart: Minor updates to the Quart app

This updates to the more recent API styles used for Quart. It is not
expected to impact the performance or validity.

* Python: Fix asyncpg PreparedStatement usage

asyncpg's PreparedStatement
(https://magicstack.github.io/asyncpg/current/api/index.html#prepared-statements)
will cache results in a local LRU cache which is not desired as per
https://github.com/TechEmpower/FrameworkBenchmarks/wiki/Project-Information-Framework-Tests-Overview#general-test-requirements
this results in some test failures whereby not enough queries are
completed.

The fix is to sample the range and hence get the `queries` number of
unique IDs rather than potentially get overlapping IDs.

I've done this for the frameworks I am familiar enough with to change
with confidence. It may apply to the other frameworks using asyncpg.

* Partially revert c18853d56f3dad20d3a72ea57cda375d26e8319a

There is a separate setup for OrJSON responses see app-orjson.py.

* Python-Starlette: Remove UJSONResponse usage

It was removed from Starlette in 0.14.1. Replaced with JSONResponse as
that is the Starlette default and hence a more realistic
representation of usage.

* Python-Sanic: Fix and upgrade sanic

* Python Starlette: Update dependencies

* Python: Utilise startup events to create DB connection (Starlette, FastAPI)

This is safer as it ensures the conneciton is created in the correct
asyncio event loop.