Complete the prototype "vertx-web-kotlinx" benchmark that produces valid results
Rename the project name from "vertx-web-kotlin" to "vertx-web-kotlinx" because it depends on some opinionated kotlinx libraries.

More performance tuning is needed. Refer to the projects "vert-web-kotlin-coroutines", "vert-web", and "vertx".

3 known exceptions during benchmarking:

1. During tests of almost all types:

   > io.netty.channel.unix.Errors$NativeIoException: recvAddress(..) failed: Connection reset by peer

2. During tests of types that require the database:

   > "logging time" io.vertx.ext.web.RoutingContext
   SEVERE: Unhandled exception in router
   io.netty.channel.StacklessClosedChannelException
   at io.netty.channel.AbstractChannel$AbstractUnsafe.write(Object, ChannelPromise)(Unknown Source)

3. During Database Updates, especially under high concurrency:

   > "logging time" io.vertx.ext.web.RoutingContext
   SEVERE: Unhandled exception in router
   vertx-web-kotlinx-postgresql: io.vertx.pgclient.PgException: ERROR: deadlock detected (40P01)

Performance results are significantly lower compared to "vertx-web-kotlin-coroutines" according to tests run on a Vagrant VirtualBox virtual machine with 24 cores and 24 GB of RAM. Most results are bwteewn 30% and 50% of corresponding ones. The result of Database Updates with `queries=20`, 24 threads, and 512 connections is only 10% of the corresponding one. The result of Plaintext is an exception, which is 2 to 3 times of the corresponding one. However, the Plaintext result in "vertx-web-kotlin-coroutines" is only about half of that in "vertx-web", which I suspect is caused by a newer version of Vert.x offering better performance on this.