Reduce the size of the toolset Docker images
The toolset Dockerfiles were using variants of `buildpack-deps` as
their base images, but this is not really necessary because we do
not compile anything (except in the wrk case); switch to the
underlying `ubuntu` images instead. Also, try to reduce the number
of layers in the images by following best practices.