H2O: Improve database query throughput
Improve the database query throughput at the cost of a slight
increase in the minimum achievable latency (e.g. if there is only
one query in progress at any given time), assuming that the
communication latency between the application server and the
database is much larger than the time it takes to perform a system
call.