Quote `randomNumber` column in PostgreSQL schema
The [test specs][] say that the `World` table has an `id` column and
`randomNumber` column (note the camelCasing).  Indeed, the
[PostgreSQL schema][] uses `randomNumber` in its `CREATE TABLE`
statement.  However, the `CREATE TABLE` statement does not quote the
column name, so PostgreSQL treats the name as if it were lowercase.

Also note that the PostgreSQL schema has `CREATE TABLE World` _and_
`CREATE TABLE "World"` statements.  Presumably, this is because the
former actually creates a `world` table because it too is not quoted.
(Whereas the latter actually creates a `World` table.)

This commit modifies the `CREATE TABLE "World"` statement to quote the
`randomNumber` column so that its camel case is preserved.

[test specs]: https://github.com/TechEmpower/FrameworkBenchmarks/wiki/Project-Information-Framework-Tests-Overview
[PostgreSQL schema]: https://github.com/TechEmpower/FrameworkBenchmarks/blob/ddd09520c10926c0e2c73a005b1f79f5a68142a8/toolset/databases/postgres/create-postgres.sql