[ruby] Use Bundler 2.6.2 (#9703)
Ruby 3.4 has Bundler 2.6.2 installed by default.
Using this version should be a bit faster as it doesn't need to install
another version as mentioned by the following warning:

      Bundler 2.6.2 is running, but your lockfile was generated with
      2.5.16. Installing Bundler 2.5.16 and restarting using that
      version.

This also outputs the platform for Rack.
If we know the platform we can improve the Ruby build times by using
precompiled binaries.