Add --force-rm flag (#9727)
* feat: add `--force-rm` flag to configure removing intermediate docker layers

This commit makes it possible to turn off the removal of intermediate
docker layers when building the tfb containers by adding a `--force-rm`
flag to the tfb script. This is useful in situations where you want to
inspect the intermediate layers for debugging purposes, or for caching
builds of dependencies as a docker layer to speed up the build process.

Note that the default behavior is to remove the intermediate layers to
avoid filling up the disk with unused layers on the citrine server.

Fixes: https://github.com/TechEmpower/FrameworkBenchmarks/issues/9718

* hyper: cache dependencies to reduce build time

This change reduces the time it takes to build the hyper Docker image by
caching the dependency builds. This is done by building a dummy binary
before copying the source code into the image.

* Change default for --force-rm to False

---------

Co-authored-by: Mike Smith <msmith@techempower.com>