Fix `org.postgresql.util.PSQLException: Transaction isolation level 0 not supported.` with https://github.com/huanshankeji/exposed-vertx-sql-client/commit/bb698565c3b7684637312a5624bc0fa6f8542196 published to Maven local, and fix some remaining issues in the benchmark code debugging with `tfb`
Here are some results with max parameters (threads, connections, number of queries) of interest as benchmarked on my device. The "single query" result is only 65% of that of `vertx-web-kotlinx`, which is not ideal. The "updates" result exceeds that of `vertx-web-kotlinx` by about 10%-15%, which needs to be looked into further. Setting `validateBatch = false` does not make noticeable difference.