Some more small performance tweaks to the `vertx-web-kotlinx` portion (#10399)
* Fix the `updates` test performance

The details of the performance issue is mentioned in the commit message of commit 6b531dd2382f6ee99194f5346be602f44bf53b19.

Changes that mostly certainly affect performance:

1. Fix capitalization and formatting in the SQL strings.
1. Use the Coroutine `awaitAll` vararg variant in `selectRandomWorlds`.

Changes that might affect performance, but I am not sure:

1. Conditionally adding only needed routes.

* Add spaces around the equal sign to make the format consistent

* Use one Kotlin `XorWowRandom` `Random` instance every `Verticle` instead of the default global one (companion object `Random`) which delegates dynamically to Java `ThreadLocalRandom` to improve some performance

Looks like there is a 1%-3% improvement in the single query test result with "32 threads and 512 connections" as tested on my device.